add pic nginx server

+++ clus_face.py
@@ -15,6 +15,11 @@ from imutils import paths
 from imutils import build_montages
 from sklearn.cluster import DBSCAN
 
+# simple log functions.
+def trace(msg):
+    date = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+    print("[%s][trace] %s"%(date, msg))
+
 def add_pic(cat_id, imagePath):
     filesize = os.path.getsize(imagePath)
     dt = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
@@ -22,7 +27,7 @@ def add_pic(cat_id, imagePath):
     try:
         image.shape
     except:
-        print('failed to read img file: %s' % (imagePath))
+        trace('failed to read img file: %s' % (imagePath))
         return False
 
     # 打开数据库连接
@@ -42,19 +47,18 @@ def add_pic(cat_id, imagePath):
         cursor.execute(sql)
         # 执行sql语句
         db.commit()
-        print('pic %s added to db.' % imagePath)
+        trace('pic %s added to db.' % imagePath)
 
         cursor.execute('SELECT id from clus_img_tb WHERE img_path = %s', imagePath)
         values = cursor.fetchall()
         img_idx = values[0][0]
-        #print('img_idx: ', img_idx)
 
         detect_face(db, cursor, imagePath, cat_id, img_idx)
         return True
     except Exception as e:
 		    # 发生错误时回滚
         db.rollback()
-        print('failed to add img record: ', e)
+        trace('failed to add img record: {}'.format(e))
     finally:
         db.close()
         
@@ -62,7 +66,6 @@ def add_pic(cat_id, imagePath):
 
 def save_face(db, cursor, imagePath, cat_id, img_idx, box, feature):
     bytes_feature = feature.tostring()
-    #print(bytes_feature)
 
     try:
       top, right, bottom, left = box
@@ -71,25 +74,17 @@ def save_face(db, cursor, imagePath, cat_id, img_idx, box, feature):
 
       db.commit()
 
-      print('img %d, face (%d, %d, %d, %d) added to db' % (img_idx, left, top, right, bottom))
-      
-      #cursor.execute('select feature from clus_face_tb where img_idx = %s' % (123))
-      #values = cursor.fetchall()
-
-      #print(values)
-      #feature = np.frombuffer(values[0][0], dtype=np.float64) #np.float32
-      #print(feature)
+      trace('img %d, face (%d, %d, %d, %d) added to db' % (img_idx, left, top, right, bottom))
     except Exception as e:
-        # 发生错误时回滚
         db.rollback()
-        print('failed to add face record: ', e)
+        trace('failed to add face record: {}'.format(e))
 
 def detect_face(db, cursor, imagePath, cat_id, img_idx):
     image = cv2.imread(imagePath)
     try:
         image.shape
     except:
-        print('failed to read img file: %s' % (imagePath))
+        trace('failed to read img file: %s' % (imagePath))
         return False
 
     rgb = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
@@ -112,10 +107,12 @@ def update_cluster(db, cursor, face_idx, label_id):
           cursor.execute(sql)
           db.commit()
       except Exception as e:
-        print('failed to update face data: ', e)
+        trace('failed to update face data: {}'.format(e))
         db.rollback()
 
 def cluster_faces(cat_id):
+    save_montage = False
+    
     # 打开数据库连接
     db = pymysql.connect("192.168.23.71","root","tysxwg07","test" )
     
@@ -131,19 +128,19 @@ def cluster_faces(cat_id):
 
       data = []
       for v in values:
-          #print(values)
+          #trace(values)
           face_idx  = v[0]
           img_idx   = v[1]
           top, right, bottom, left = v[2:6]
           feature   = np.frombuffer(v[6], dtype=np.float64)
           img_path  = v[7]
 
-          #print(feature)
+          #trace(feature)
           box = (top, right, bottom, left)
           d = [{"faceIndex": face_idx, "imageIndex": img_idx, "imagePath": img_path, "loc": box, "encoding": feature}]
           data.extend(d)
     except Exception as e:
-        print('failed to read face data: ', e)
+        trace('failed to read face data: {}'.format(e))
         db.close()
         return
 
@@ -151,29 +148,28 @@ def cluster_faces(cat_id):
     encodings = np.array([d["encoding"] for d in data])
     #np.random.shuffle(encodings)
 
-    print(encodings.shape)
-    #print(encodings[0])
-    #print(type(encodings))
+    trace(encodings.shape)
 
     # cluster the embeddings
-    print("[INFO] clustering...")
+    trace("[INFO] clustering...")
     clt = DBSCAN(metric="euclidean", n_jobs=-1, eps=0.35, min_samples=3)
     clt.fit(encodings)
 
-    plt.scatter(encodings[:, 4], encodings[:, 9], marker='o', c=clt.labels_)
-    plt.show()
+    #plt.scatter(encodings[:, 4], encodings[:, 9], marker='o', c=clt.labels_)
+    #plt.show()
 
     # determine the total number of unique faces found in the dataset
     labelIDs = np.unique(clt.labels_)
     numUniqueFaces = len(np.where(labelIDs > -1)[0])
-    print("[INFO] # unique faces : {}".format(numUniqueFaces))
+    trace("[INFO] # unique faces : {}".format(numUniqueFaces))
 
-    # remove old pic
-    for num in range(0, 20):
-        filename = 'out_%d.jpg' % num
-        if os.path.exists(filename):
-          os.remove(filename)
-          print('file %s deleted' % filename)
+    if save_montage:
+        # remove old pic
+        for num in range(0, 20):
+            filename = 'out_%d.jpg' % num
+            if os.path.exists(filename):
+              os.remove(filename)
+              trace('file %s deleted' % filename)
       
     # loop over the unique face integers
     face_id = 0
@@ -181,44 +177,45 @@ def cluster_faces(cat_id):
     for labelID in labelIDs:
         # find all the indexes into the 'data' array that belong to the
         # current label ID, then randomly sample a maximum of 25 index from the set
-        print("[INFO] faces for face ID: {}".format(labelID))
+        trace("[INFO] faces for face ID: {}".format(labelID))
         idxs = np.where(clt.labels_ == labelID)[0]
-        print("[INFO] face count: %d, cluster count %d" % (len(idxs), cluster_face_count))
+        trace("[INFO] face count: %d, cluster count %d" % (len(idxs), cluster_face_count))
         cluster_face_count = cluster_face_count + len(idxs)
         
         if labelID != -1:
             [update_cluster(db, cursor, data[i]["faceIndex"], labelID) for i in idxs]
 
-        idxs = np.random.choice(idxs, size=min(25, len(idxs)),
-          replace=False)
+        if save_montage:
+            idxs = np.random.choice(idxs, size=min(25, len(idxs)),
+              replace=False)
 
-        # initialize the list of faces to include in the montage
-        faces = []
-        # loop over the sampled indexes
-        for i in idxs:
-            # load the input image and extract the face ROI
-            image = cv2.imread(data[i]["imagePath"])
-            (top, right, bottom, left) = data[i]["loc"]
-            face = image[top:bottom, left:right]
+            # initialize the list of faces to include in the montage
+            faces = []
+            # loop over the sampled indexes
+            for i in idxs:
+                # load the input image and extract the face ROI
+                image = cv2.imread(data[i]["imagePath"])
+                (top, right, bottom, left) = data[i]["loc"]
+                face = image[top:bottom, left:right]
 
-            # force resize the face ROI to 96x96 and then add it to the
-            # faces montage list
-            face = cv2.resize(face, (96,96))
-            faces.append(face)
+                # force resize the face ROI to 96x96 and then add it to the
+                # faces montage list
+                face = cv2.resize(face, (96,96))
+                faces.append(face)
 
-        # create a montage using 96x96 "tiles" with 5 rows and 5 columns
-        montage = build_montages(faces, (96,96), (5,5))[0]
+            # create a montage using 96x96 "tiles" with 5 rows and 5 columns
+            montage = build_montages(faces, (96,96), (5,5))[0]
 
-        # show the output montage
-        title = "Face ID #{}".format(labelID)
-        title = "Unknown Faces" if labelID == -1 else title
-        #cv2.imshow(title, montage)
-        #cv2.waitKey(0)
-        cv2.imwrite('out_{}.jpg'.format(face_id), montage)
+            # show the output montage
+            title = "Face ID #{}".format(labelID)
+            title = "Unknown Faces" if labelID == -1 else title
+            #cv2.imshow(title, montage)
+            #cv2.waitKey(0)
+            cv2.imwrite('out_{}.jpg'.format(face_id), montage)
         face_id += 1
 
     db.close()
-    print("[INFO] total cluster_face_count %d" % cluster_face_count)
+    trace("[INFO] total cluster_face_count %d" % cluster_face_count)
 
 def test():
     cat_id = 'test'
@@ -228,12 +225,19 @@ def test():
     # 关闭数据库连接
     db.close()
     
+def get_msec():
+    msec = int(time.time() * 1000)
+    return msec
+    
 if __name__ == '__main__':
+    total_cluster_msec = 0
     while True:
+        trace('ready to scan pic')
+        t = int(time.time())
         for folder in os.listdir(config.upload_folder):
             if os.path.isdir(os.path.join(config.upload_folder, folder)):
                 cat_id = folder
-                print('list pics in cat_id: %s' % cat_id)
+                trace('list pics in cat_id: %s' % cat_id)
                 
                 cluster_face = False
                 for file in os.listdir(os.path.join(config.upload_folder, folder)):
@@ -249,13 +253,21 @@ if __name__ == '__main__':
                         if add_pic(cat_id, new_filepath):
                             cluster_face = True
                         else:
-                            print('failed to read pic')
+                            trace('failed to read pic')
                             shutil.remove(new_filepath)
                     except Exception as e:
-                        print('failed to move file: ', e)
+                        trace('failed to move file: {}'.format(e))
                 
                 if cluster_face:
-                    print('cluster face: %s' % cat_id)
+                    trace('cluster face: %s' % cat_id)
+                    start_time = get_msec()
                     cluster_faces(cat_id)
-        time.sleep(30)
+                    cluster_msec = get_msec() - start_time
+                    trace('cluster_msec: %d (avg %d)' % (cluster_msec, total_cluster_msec))
+                    total_cluster_msec = (total_cluster_msec * 4 + cluster_msec) // 5
+                    
+        used_time = int(time.time()) - t
+        sleep_sec = 30 - used_time
+        if sleep_sec > 3:
+            time.sleep(sleep_sec)
